# runtime: add ERMS-based memmove support for modern CPU platforms